speculation-rules/prefetch/anonymous-client.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="../resources/utils.js"></script>
<script src="resources/utils.sub.js"></script>
<script src="/resources/testdriver.js?feature=bidi"></script>
<script src="/resources/testdriver-extra.js"></script>
<script>
  setup(() => assertSpeculationRulesIsSupported());

  promise_test(async t => {
    let agent = await spawnWindow(t);
    let nextUrl = agent.getExecutorURL({ hostname: CROSS_ORIGIN_HOST_THAT_WORKS_WITH_ACIWCO, page: 2 });

    await test_driver.bidi.speculation.prefetch_status_updated.subscribe();
  let statusPromise = new Promise(resolve => {
    test_driver.bidi.speculation.prefetch_status_updated.on(event => {
      if (event && event.status !== undefined) resolve(event.status);
    });
  });

    await agent.forceSinglePrefetch(nextUrl, { requires: ["anonymous-client-ip-when-cross-origin"] });
    await agent.navigate(nextUrl);

    let requestHeaders = await agent.getRequestHeaders();
    assert_prefetched_anonymous_client_ip(requestHeaders);
  }, "test anonymous-client url prefetch for cross origin pages");
</script>
